% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBconnect.r
\name{DBconnect}
\alias{DBconnect}
\title{Connect to Access database}
\usage{
DBconnect(databases, db, dir)
}
\arguments{
\item{databases}{list of valid databases}

\item{db}{name code for database; see databases.txt for a list of codes}

\item{dir}{directory location for databases; if NULL uses value in databases.txt; if "" uses package directory}
}
\value{
connection if all good
}
\description{
Function used by getCalcurData and saveCalcurData to connect to Access databases.
}
\author{
Jeff Laake
}
