% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.conversion.r
\name{ejcheck}
\alias{ejcheck}
\title{Validates Ej resight data}
\usage{
ejcheck(filename = NULL)
}
\arguments{
\item{filename}{name of file output from ejresights function; if NULL can select with file browser}
}
\value{
no return value
}
\description{
Validates Ej resight data
}
\author{
Jeff Laake, Jeff Harris
}
