% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.conversion.r
\name{ejresights}
\alias{ejresights}
\title{Convert resight file from ACDSee format for Ej sightings to data entry template format}
\usage{
ejresights(inp = NULL, out, append = FALSE)
}
\arguments{
\item{inp}{name of input file containing metadata from ACDsee; if NULL can select from file browser}

\item{out}{name of output file which is formatted for data copying to data entry template}

\item{append}{if TRUE, append to the output file}
}
\value{
no return value
}
\description{
Convert resight file from ACDSee format for Ej sightings to data entry template format
}
\author{
Jeff Laake, Jeff Harris
}

