% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buoy.r
\name{read_ndbc}
\alias{read_ndbc}
\alias{read_ndbc_month}
\title{Retrieve buoy data}
\usage{
read_ndbc(buoy, year)
}
\arguments{
\item{year}{4 digit year to extract}

\item{bouy}{NDBC buoy designation (ie ESB 46053)}

\item{month}{text string for month from current year data}

\item{dir}{directory location for databases; if NULL uses value in databases.txt; if "" uses package directory}
}
\value{
extracted dataframe
}
\description{
Retrieve buoy data
}
\examples{
esb_2013=read_ndbc("46053",2013)
esb_jan=read_ndbc_month("46053","Jan",2014)
}
\author{
Jeff Laake
}

