% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.conversion.r
\name{resights}
\alias{resights}
\title{Convert resight file from ACDSee format to data entry template format}
\usage{
resights(inp, out, append = FALSE)
}
\arguments{
\item{inp}{name of input file containing metadata from ACDsee}

\item{out}{name of output file which is formatted for data copying to data entry template}

\item{append}{if TRUE, append to the output file}
}
\value{
no return value
}
\description{
Convert resight file from ACDSee format to data entry template format
}
\author{
Jeff Laake, Jeff Harris
}
