% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCalcurData.r
\name{saveCalcurData}
\alias{saveCalcurData}
\title{Retrieve tables from Calcur Databases}
\usage{
saveCalcurData(x, db, tbl, dir = NULL)
}
\arguments{
\item{x}{dataframe to be saved as table}

\item{db}{name code for database; see databases.txt for a list of codes}

\item{tbl}{name of table in database}

\item{dir}{directory location for databases; if NULL uses files with package}
}
\description{
This function acts as an interface for other software to save tables
in the databases in a way that keeps the other functions independent from
changes to the way the data are stored.  However, the names of the tables
and the basic table structure within each database must remain static. It only
works with 32 bit R if Access is 32 bit.  It is written with this presumption and
if there is a connection error then it will tell you to try 32 bit R if the version
being used is 64 bit.  If 64 bit Access is available it presumably will only work
with 64 bit R because the 32 and 64 bit ODBC implementations cannot be on the same machine
presently.  All access to the tables is accomplished with the R package RODBC.
}
\details{
There is only one way to run this function:
1) saveCalcurData(x, db="Zc", tbl="Alive") will save the contents in x to table tbl in the specified 
dataframe after removing any existing table with that name.  Names here are case-specific unlike
with getCalcurData.

The default directory specification is the one in databases.txt.  If you specify dir="" then
it will expect to find the databases in the directory for the package location in the 
library.  I use that for testing and to have a local copy not on the network.
Any location can be used as long as the database file is at that location.
}
\author{
Jeff Laake
}
