% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_buoy.r
\name{update_buoy_data}
\alias{update_buoy_data}
\alias{update_all}
\title{Update buoy data}
\usage{
update_all(month=NULL,year=NULL,dir=NULL)
       update_buoy_data(newdata,tablename,dir=NULL)
}
\arguments{
\item{newdata}{newly extracted buoy data to add to database}

\item{tablename}{name of table in environmental database (eg "PtSanLuisDailySST")}

\item{dir}{directory location for databases; if NULL uses value in databases.txt; if "" uses package directory}

\item{month}{text string for month from current year data}

\item{year}{4 digit year to extract}
}
\value{
no return value
}
\description{
update_buoy_data adds data to existing table but deletes records in common in original data
update_all uses update_buoy_data to add a month or year to each of the existing tables
}
\examples{
esb_2013=read_ndbc("46053",2013)
# commented out so has not to modify data
#update_buoy_data(esb_2013,"EastSantaBarbaraChannelBuoyData")
#update_all(month="Jan",2014)
}
\author{
Jeff Laake
}
