% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_wtmp}
\alias{construct_daily}
\alias{filter_wtmp}
\title{Utility functions}
\usage{
filter_wtmp(x)
}
\arguments{
\item{x}{dataframe of bouy data}

\item{data}{dataframe of buoy data}
}
\value{
manipulated dataframe
}
\description{
filter_wtmp throws out records in which water temp is unknown 999
construct_daily constructs daily averages; this is used with PtSan Luis because original data was only a single value per day
}
\author{
Jeff Laake
}

