% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{filter_wtmp}
\alias{construct_daily}
\alias{filter_wtmp}
\title{Utility functions}
\usage{
filter_wtmp(x)
       construct_daily(data,buoy="PSLC1")
}
\arguments{
\item{x}{dataframe of bouy data}

\item{data}{dataframe of buoy data}

\item{buoy}{character designation for the buoy}
}
\value{
manipulated dataframe
}
\description{
filter_wtmp throws out records in which water temp is unknown 999;
construct_daily constructs daily averages; this is used with PtSan Luis because original data was only a single value per day
}
\author{
Jeff Laake
}

