% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.r
\name{read_ndbc}
\alias{read_ndbc}
\alias{read_ndbc_month}
\title{Retrieve buoy data}
\usage{
read_ndbc(buoy,year)
       read_ndbc_month(buoy,month,year)
}
\arguments{
\item{buoy}{NDBC buoy designation (ie ESB 46053)}

\item{year}{4 digit year to extract}

\item{month}{text string for month from current year data}
}
\value{
extracted dataframe
}
\description{
Retrieve buoy data
}
\examples{
esb_2013=read_ndbc("46053",2013)
esb_jan=read_ndbc_month("46053","Jan",2014)
}
\author{
Jeff Laake
}
